requirements.txt
google-generativeai
fastapi
uvicorn
google-generativeai
hyperon
python-dotenv